refactor: decouple generated asset classes by introducing flutter_gen_interface package
