+++ packages/flutter_gen_interface/.metadata
@@ -1,0 +1,10 @@
+# This file tracks properties of this Flutter project.
+# Used by Flutter tool to assess capabilities and perform upgrades etc.
+#
+# This file should be version controlled and should not be manually edited.
+
+version:
+  revision: "78d2ce61029f9cd4271f0bcca8a6a178c59455d3"
+  channel: "[user-branch]"
+
+project_type: package
