+++ packages/flutter_gen_interface/CHANGELOG.md
@@ -1,0 +1,1 @@
+../../CHANGELOG.md
