+++ packages/flutter_gen_interface/LICENSE
@@ -1,0 +1,1 @@
+../../LICENSE
