+++ packages/flutter_gen_interface/analysis_options.yaml
@@ -1,0 +1,1 @@
+include: ../../analysis_options.yaml
